+++ projects/ngx-scrollbar/karma.conf.js
@@ -30,7 +30,8 @@ module.exports = function (config) {
       reporters: [
         { type: 'html' },
         { type: 'text-summary' },
-        { type: 'cobertura' }
+        { type: 'cobertura' },
+        { type: 'lcov' }
       ]
     },
     reporters: ['progress', 'kjhtml'],