+++ projects/ngx-scrollbar/src/lib/tests/fit.spec.ts
@@ -1,5 +1,6 @@
 import { ComponentFixture, ComponentFixtureAutoDetect, TestBed } from '@angular/core/testing';
 import { Directionality } from '@angular/cdk/bidi';
+import { By } from '@angular/platform-browser';
 import { BehaviorSubject, firstValueFrom } from 'rxjs';
 import { NgScrollbar } from 'ngx-scrollbar';
 import { setDimensions } from './common-test.';
@@ -24,9 +25,12 @@ describe('Fit styles', () => {
     directionalityMock.value = 'ltr';
     fixture = TestBed.createComponent(NgScrollbar);
     component = fixture.componentInstance;
+
+    // Set scrollbar offset value
+    component.nativeElement.style.setProperty('--scrollbar-offset', '4');
   });
 
-  let scrollbarOffset: number = 8;
+  const scrollbarOffset: number = 8;
 
   it('should fit both _scrollbars only if both of them are displayed', async () => {
     setDimensions(component, { cmpHeight: 200, cmpWidth: 200, contentHeight: 500, contentWidth: 500 });
@@ -34,11 +38,11 @@ describe('Fit styles', () => {
     component.ngAfterViewInit();
     await firstValueFrom(component.afterInit);
 
-    const trackYElement: HTMLElement = component._scrollbars.y.track.nativeElement;
-    const trackXElement: Element = component._scrollbars.x.track.nativeElement;
+    const trackYElement: Element = fixture.debugElement.query(By.css('scrollbar-y .ng-scrollbar-track')).nativeElement;
+    const trackXElement: Element = fixture.debugElement.query(By.css('scrollbar-x .ng-scrollbar-track')).nativeElement;
 
-    const stickyYElement: Element = component._scrollbars.y.sticky.nativeElement;
-    const stickyXElement: Element = component._scrollbars.x.sticky.nativeElement;
+    const stickyYElement: Element = fixture.debugElement.query(By.css('scrollbar-y .ng-scrollbar-sticky')).nativeElement;
+    const stickyXElement: Element = fixture.debugElement.query(By.css('scrollbar-x .ng-scrollbar-sticky')).nativeElement;
 
     expect(trackXElement.clientWidth).toBe(200 - stickyYElement.clientWidth - scrollbarOffset);
     expect(trackYElement.clientHeight).toBe(200 - stickyXElement.clientHeight - scrollbarOffset);
@@ -50,7 +54,8 @@ describe('Fit styles', () => {
     component.ngAfterViewInit();
     await firstValueFrom(component.afterInit);
 
-    const trackYElement: Element = component._scrollbars.y.track.nativeElement;
+    const trackYElement: Element = fixture.debugElement.query(By.css('scrollbar-y .ng-scrollbar-track')).nativeElement;
+
     expect(trackYElement.clientHeight).toBe(200 - scrollbarOffset);
   });
 
@@ -60,7 +65,7 @@ describe('Fit styles', () => {
     component.ngAfterViewInit();
     await firstValueFrom(component.afterInit);
 
-    const trackXElement: Element = component._scrollbars.x.track.nativeElement;
+    const trackXElement: Element = fixture.debugElement.query(By.css('scrollbar-x .ng-scrollbar-track')).nativeElement;
     expect(trackXElement.clientWidth).toBe(200 - scrollbarOffset);
   });
 });