+++ projects/ngx-scrollbar/src/lib/utils/scrollbar-manager.ts
@@ -1,7 +1,7 @@
 import { Injectable, inject, signal, WritableSignal, PLATFORM_ID } from '@angular/core';
 import { DOCUMENT, isPlatformBrowser } from '@angular/common';
-import { getRtlScrollAxisType, RtlScrollAxisType } from '@angular/cdk/platform';
 import { NG_SCROLLBAR_POLYFILL } from '../ng-scrollbar.model';
+import { ScrollTimelineFunc } from './common';
 
 const scrollTimelinePolyfillUrl: string = 'https://flackr.github.io/scroll-timeline/dist/scroll-timeline.js';
 
@@ -16,12 +16,7 @@ export class ScrollbarManager {
 
   readonly window: Window = this.document.defaultView;
 
-  /**
-   * Indicates the RTL scrollAxisType (used for dragging functionality)
-   */
-  readonly rtlScrollAxisType: RtlScrollAxisType = getRtlScrollAxisType();
-
-  readonly scrollTimelinePolyfill: WritableSignal<any> = signal(null);
+  readonly scrollTimelinePolyfill: WritableSignal<ScrollTimelineFunc> = signal(null);
 
   constructor() {
     if (this.isBrowser && (!this.window['ScrollTimeline'] || !CSS.supports('animation-timeline', 'scroll()'))) {
@@ -36,7 +31,7 @@ export class ScrollbarManager {
       script.src = this._polyfillUrl;
 
       // Wait for the script to load
-      await new Promise<any>((resolve, reject) => {
+      await new Promise<Event>((resolve, reject) => {
         script.onload = resolve;
         script.onerror = reject;
         this.document.head.appendChild(script);